+++ src/dpe/dpe_worker_node.h
@@ -42,7 +42,7 @@ public:
   
   int handleRequest(const Request& req, Response& reply);
   
-  void removeNode(int id);
+  void removeNode(int64 id);
 private:
   WorkerTaskExecuter taskExecuter;
   RemoteNodeImpl* remoteNode;
